Move pane lifecycle observers and service construction into the dock pane
MVC separation: the dock pane (a HasTraits, not a Qt widget) now owns the
RowManager / sync / sticky / experiment-manager / state-tracker defaults
and the @observe wiring for manager events, experiment changes, and the
exit veto. This retires the pane-side manual observe()/closeEvent
detach dance entirely - Traits manages the observer lifecycle, and no
application observer can dispatch into a deleted Qt widget.

Refactor fixes on top of the move:
- protocol_state_tracker and sticky_manager gained defaults (they were
  None, crashing the manager observers and silently disabling sticky
  notes); the tracker default re-binds dock_pane=self so the
  [modified] title keeps updating.
- Guard self.control access in observers that can fire before
  create_contents.
- Restore the application-is-None guard for the initial experiment
  label (standalone demos construct the pane without an application).
- Drop the pane imports that moved with the reconciliation helpers and
  the dead _log_ctx None-check; the logging device context is rebuilt
  inline on every run start.
- Test fixture passes task at construction (traits_init now seeds the
  ack-wait preferences, matching production where the dock-pane factory
  receives task=task).
