+++ pluggable_protocol_tree/views/dock_pane.py
@@ -3,20 +3,25 @@
 Receives its column set from the plugin on construction and constructs
 the experiment + sticky-note services from the live Envisage
 application so the experiment-bar buttons drive real handlers."""
-
+from microdrop_application.dialogs.pyface_wrapper import confirm, YES
+from pluggable_protocol_tree.consts import REPEAT_DURATION_RECALC_TRIGGERS
+from pluggable_protocol_tree.services.phase_math import effective_repetitions_for_duration, estimate_repeat_duration_s
+from pluggable_protocol_tree.services.protocol_state_tracker import PluggableProtocolStateTracker
 from pyface.tasks.api import TraitsDockPane
-from traits.api import Instance, List, Str
+from traits.api import Instance, List, Str, observe
 
-from device_viewer.consts import CHANNEL_AREAS_KEY, DEVICE_SVG_PATH_KEY
 from logger.logger_service import get_logger
 from microdrop_application.helpers import get_microdrop_redis_globals_manager
 from microdrop_utils.sticky_notes import StickyWindowManager
+from pluggable_protocol_tree.models.row_manager import RowManager
+from pluggable_protocol_tree.services.device_viewer_sync import DeviceViewerSyncController
+from pluggable_protocol_tree.views.protocol_tree_pane import ProtocolTreePane, REPEAT_DURATION_TOLERANCE_S, \
+    REPEAT_DURATION_DECIMALS
 from protocol_grid.services.experiment_manager import ExperimentManager
 
 from pluggable_protocol_tree.interfaces.i_column import IColumn
-from pluggable_protocol_tree.services.logging.models import LoggingDeviceContext
 from pluggable_protocol_tree.services.preferences import (
-    ProtocolPreferences, seed_ack_times,
+    ProtocolPreferences
 )
 
 logger = get_logger(__name__)
@@ -32,7 +37,12 @@ class PluggableProtocolDockPane(TraitsDockPane):
     name = Str("Protocol (pluggable)")
 
     columns = List(Instance(IColumn))
+    manager = Instance(RowManager)
+    sync = Instance(DeviceViewerSyncController)
+    sticky_manager = Instance(StickyWindowManager)
+    experiment_manager = Instance(ExperimentManager)
     quick_actions = List(desc="Quick actions to mount under the tree.")
+    protocol_state_tracker = Instance(PluggableProtocolStateTracker)
 
     #: Protocol preferences model (the "microdrop.protocol" node). Bound to
     #: the live application's preferences in create_contents, then passed
@@ -41,63 +51,45 @@ class PluggableProtocolDockPane(TraitsDockPane):
     #: visibility).
     preferences = Instance(ProtocolPreferences)
 
-    def create_contents(self, parent):
-        # Local imports to avoid pulling Qt at plugin-import time.
-        from pluggable_protocol_tree.models.row_manager import RowManager
-        from pluggable_protocol_tree.services.device_viewer_sync import (
-            DeviceViewerSyncController,
-        )
-        from pluggable_protocol_tree.views.protocol_tree_pane import (
-            ProtocolTreePane,
-        )
+    def _experiment_manager_default(self):
+        return ExperimentManager(self.task.window.application.current_experiment_directory)
 
-        app = self.task.window.application
-        if self.preferences is None:
-            self.preferences = ProtocolPreferences(preferences=app.preferences)
+    def _sticky_manager_default(self):
+        return StickyWindowManager()
 
+    def _protocol_state_tracker_default(self):
+        # dock_pane=self binds the tracker's display name ("<name> -
+        # <protocol> [modified]") to this pane's title.
+        return PluggableProtocolStateTracker(dock_pane=self)
+
+    def _preferences_default(self):
+        return ProtocolPreferences(preferences=self.task.window.application.preferences)
+
+    def _sync_default(self):
+        return DeviceViewerSyncController(row_manager=self.manager)
+
+    def _manager_default(self):
+        return RowManager(columns=list(self.columns))
+
+    def traits_init(self):
         # One ack-wait grid entry per wait-capable column, the plugin
         # provider's default_ack_time_s as the wait time; user-edited
         # values persisted on the node are kept.
-        seed_ack_times(self.preferences, self.columns)
+        self.preferences.seed_ack_times_from_columns(self.columns)
 
-        experiment_manager = ExperimentManager(app.current_experiment_directory)
-        sticky_manager = StickyWindowManager()
-
-        manager = RowManager(columns=list(self.columns))
-        sync = DeviceViewerSyncController(row_manager=manager)
-        def _logging_device_context():
-            # Decoupled: read channel areas + device SVG path from the shared
-            # app_globals (published by the device viewer) rather than reaching
-            # into the device-viewer pane/model.
-            channel_areas, svg_path = {}, None
-            try:
-                # Redis JSON-stringifies the int channel keys; restore int keys
-                # to match the actuation lookup (areas.get(int(ch))).
-                channel_areas = {
-                    int(k): float(v)
-                    for k, v in (app_globals.get(CHANNEL_AREAS_KEY) or {}).items()
-                }
-                svg_path = app_globals.get(DEVICE_SVG_PATH_KEY)
-            except Exception as e:
-                logger.debug(f"logging device-context probe failed: {e}")
-            return LoggingDeviceContext(
-                experiment_directory=experiment_manager.get_experiment_directory(),
-                device_svg_path=svg_path,
-                channel_areas=channel_areas,
-            )
-
+    def create_contents(self, parent):
         pane = ProtocolTreePane(
-            manager,
-            application=app,
-            experiment_manager=experiment_manager,
-            sticky_manager=sticky_manager,
-            device_viewer_sync=sync,
-            logging_device_context_provider=_logging_device_context,
+            self.manager,
+            application=self.task.window.application,
+            experiment_manager=self.experiment_manager,
+            sticky_manager=self.sticky_manager,
+            device_viewer_sync=self.sync,
             preferences=self.preferences,
             quick_actions=list(self.quick_actions),
+            protocol_state_tracker=self.protocol_state_tracker,
             parent=parent,
         )
-        pane.protocol_state_tracker.dock_pane = self
+
         # Legacy protocol_grid parity: the full app opens with one default
         # step when no protocol is loaded (no-op once a protocol is loaded).
         pane._seed_default_step_if_empty()
@@ -124,3 +116,142 @@ class PluggableProtocolDockPane(TraitsDockPane):
         # Reuses the same handler the experiment-bar button drives so
         # the menu and the toolbutton stay consistent.
         self._pane()._on_new_experiment()
+
+    ### Trait observers ###########################
+    @observe("manager.rows_changed")
+    def _on_manager_rows_changed(self, event):
+        """Structural mutation — re-check the baseline path set."""
+        self.protocol_state_tracker.on_structure_changed(self.manager)
+
+    @observe("manager.cell_changed")
+    def _on_manager_cell_changed(self, event):
+        """Cell value edit — incremental dirty update for the one cell."""
+        payload = event.new
+
+        if not isinstance(payload, dict):
+            return
+
+        path = payload.get("path")
+        col_id = payload.get("col_id")
+
+        if path is None or col_id is None:
+            return
+
+        self.protocol_state_tracker.on_cell_changed(
+            path, col_id, self.manager,
+        )
+
+        self._clamp_trail_overlay_for_row(path, col_id)
+        self._reconcile_repeat_duration_for_row(path, col_id)
+
+    @observe("task.window.application.experiment_changed")
+    def _on_experiment_changed(self, event):
+        # control is None until create_contents has run (the application
+        # can switch experiments before this pane is mounted).
+        if self.control is None:
+            return
+        self.control.widget()._on_experiment_changed()
+
+    @observe("task.window.application.application_exiting")
+    def _on_application_exiting(self, event):
+        """Veto application exit when the protocol is dirty and the user
+        elects to keep it open.
+
+        ``event`` is a Pyface Vetoable event — setting ``event.veto = True``
+        cancels the exit.
+        """
+        if not self.protocol_state_tracker.is_modified:
+            return
+        user_choice = confirm(
+            self,
+            "Current protocol has unsaved changes.\n"
+            "Exit without saving?",
+            title="Unsaved Protocol Changes",
+            cancel=False,
+        )
+        if user_choice != YES:
+            event.veto = True
+
+    ######### Helpers ###################
+    def _clamp_trail_overlay_for_row(self, path, col_id):
+        """Mirror the DV sidebar's dynamic bound (trail_overlay can never
+        reach trail_length): shrinking Trail Len drags an out-of-range
+        Trail Overlay down with it. Runs before the repeat-duration
+        reconciliation so the recalc sees the clamped overlay."""
+        if col_id != "trail_length":
+            return
+        try:
+            row = self.manager.get_row(tuple(path))
+        except (IndexError, AttributeError):
+            return
+        max_overlay = max(0, int(getattr(row, "trail_length", 1) or 1) - 1)
+        if int(getattr(row, "trail_overlay", 0) or 0) > max_overlay:
+            row.trail_overlay = max_overlay
+            self.manager.cell_changed = {
+                "path": tuple(path), "col_id": "trail_overlay",
+            }
+
+    def _reconcile_repeat_duration_for_row(self, path, col_id):
+        """Mirror the legacy auto-recalc / effective-reps coupling:
+
+          * In Route-Reps-controlled mode (``repeat_duration_controls``
+            False): edits to any geometry/timing knob refresh the
+            Route Reps Dur cell with the new estimate.
+          * In Route-Reps-Dur-controlled mode (flag True): edits to
+            Route Reps Dur refresh the Route Reps cell with the effective
+            number of full cycles that fit.
+
+        Programmatic writes here go via ``setattr`` directly (NOT
+        ``model.set_value`` and NOT through ``on_interact``) so the
+        mode-switch dialog only ever fires for genuine user clicks,
+        never for these reconciliation passes.
+        """
+        if self.control is None or self.control.widget()._is_protocol_active():
+            return
+        try:
+            row = self.manager.get_row(tuple(path))
+        except (IndexError, AttributeError):
+            return
+        routes = list(getattr(row, "routes", []) or [])
+        if not routes:
+            return
+        controls = bool(getattr(row, "repeat_duration_controls", False))
+        duration_s = float(getattr(row, "duration_s", 1.0) or 0.0)
+        trail_length = int(getattr(row, "trail_length", 1) or 1)
+        trail_overlay = int(getattr(row, "trail_overlay", 0) or 0)
+        linear_repeats = bool(getattr(row, "linear_repeats", False))
+        soft_start = bool(getattr(row, "soft_start", False))
+        soft_end = bool(getattr(row, "soft_end", False))
+
+        if not controls and col_id in REPEAT_DURATION_RECALC_TRIGGERS:
+            n_repeats = int(getattr(row, "route_repetitions", 1) or 1)
+            estimated = estimate_repeat_duration_s(
+                routes=routes,
+                trail_length=trail_length, trail_overlay=trail_overlay,
+                n_repeats=n_repeats, step_duration_s=duration_s,
+                linear_repeats=linear_repeats,
+                soft_start=soft_start, soft_end=soft_end,
+            )
+            estimated = round(estimated, REPEAT_DURATION_DECIMALS)
+            if (abs(float(getattr(row, "repeat_duration", 0.0)) - estimated)
+                    >= REPEAT_DURATION_TOLERANCE_S):
+                row.repeat_duration = estimated
+                # Re-entrancy is bounded: see the
+                # REPEAT_DURATION_RECALC_TRIGGERS guard + mode-check above;
+                # "repeat_duration" is not a trigger in
+                # route-reps-controlled mode so the next pass exits cleanly.
+                self.manager.cell_changed = {
+                    "path": tuple(path), "col_id": "repeat_duration",
+                }
+        elif controls and col_id == "repeat_duration":
+            effective = effective_repetitions_for_duration(
+                routes=routes,
+                trail_length=trail_length, trail_overlay=trail_overlay,
+                step_duration_s=duration_s,
+                repeat_duration_s=float(getattr(row, "repeat_duration", 0.0) or 0.0),
+            )
+            if int(getattr(row, "route_repetitions", 1) or 1) != int(effective):
+                row.route_repetitions = int(effective)
+                self.manager.cell_changed = {
+                    "path": tuple(path), "col_id": "route_repetitions",
+                }
